+++ code/system.ui/react.dev/src/Dev.mts
@@ -1,7 +1,7 @@
 import { Spec } from './Spec';
 import { Harness } from './ui/Harness';
 import { Entry } from './ui/Entry';
-import { DevBus as Bus } from './ui.Bus';
+import { DevBus as Bus } from './logic.Bus';
 
 const { render } = Entry;
