+++ code/system.ui/react.dev/src/types.mts
@@ -1,7 +1,7 @@
 export * from './Spec/types.mjs';
 export * from './Spec.Context/types.mjs';
 
-export * from './ui.Bus/types.mjs';
+export * from './logic.Bus/types.mjs';
 
 export * from './ui/Entry/types.mjs';
 export * from './ui/Entry.SpecList/types.mjs';
