+++ code/system.ui/react.dev/src/ui.Hooks/common.ts
@@ -1,2 +1,2 @@
 export * from '../common';
-export { DevBus } from '../ui.Bus';
+export { DevBus } from '../logic.Bus';
