+++ code/system.ui/react.dev/src/ui/common.ts
@@ -1,4 +1,4 @@
 export * from '../common';
-export * from '../ui.Bus';
+export * from '../logic.Bus';
 export * from '../ui.Hooks';
 export { Spec, SpecContext } from '../Spec';
